Split lib/dnssec.ts into per-concern modules
Break the 960-line validator (and its 903-line test file) into a
lib/dnssec/ folder: types, algorithm policy, canonical wire encoding,
crypto plumbing, DS linkage, RRSIG verification, RRset classification,
and the chain walk -- each with its own test file. index.ts re-exports
the same public API, so no call sites change. No behavior changes.
