+++ lib/dnssec/chain.test.ts
@@ -1,0 +1,223 @@
+import type { DnskeyData, DsData } from 'dns-packet';
+import { describe, expect, it } from 'vitest';
+
+import { buildChain, ROOT_TRUST_ANCHORS } from './chain';
+import { dsDigest } from './ds';
+import { dsForKey, genKey, signDnskeyRrset } from './test-helpers';
+import { RFC_DS, RFC_KEY } from './test-vectors';
+import type { RawZone } from './types';
+import { computeKeyTag, dnskeyRdata } from './wire';
+
+describe('buildChain', () => {
+  const sep: DnskeyData = { ...RFC_KEY, flags: 257 }; // SEP/KSK bit set
+
+  // DS records that authenticate `sep` at a given owner name. The first zone in
+  // these chains is non-root, so its dsRecords act as the (trusted) parent DS --
+  // letting us exercise the secure path without forging the real root KSK.
+  const dsFor = (name: string): DsData => ({
+    ...RFC_DS,
+    keyTag: computeKeyTag(dnskeyRdata(sep)),
+    digest: dsDigest(name, sep, 1)!,
+  });
+
+  it('marks a fully linked, validly-signed chain secure', () => {
+    // Real signatures required now: each zone's DNSKEY RRset must carry a valid
+    // RRSIG by a DS-linked key, so we sign with keypairs we control.
+    const parent = genKey(13);
+    const child = genKey(13);
+    const win = { inception: 1000, expiration: 2000 };
+    const zones: RawZone[] = [
+      {
+        name: 'example',
+        keys: [parent.dnskey],
+        dsRecords: [dsForKey('example', parent.dnskey)],
+        keyRrsigs: [signDnskeyRrset('example', [parent.dnskey], parent, win)],
+      },
+      {
+        name: 'child.example',
+        keys: [child.dnskey],
+        dsRecords: [dsForKey('child.example', child.dnskey)],
+        keyRrsigs: [
+          signDnskeyRrset('child.example', [child.dnskey], child, win),
+        ],
+      },
+    ];
+    const chain = buildChain(zones, 1500);
+    expect(chain.zones.map((z) => z.status)).toEqual(['secure', 'secure']);
+    expect(chain.overall).toBe('secure');
+  });
+
+  it('enforces the real root trust anchor (fake root key -> broken)', () => {
+    const chain = buildChain([{ name: '.', keys: [sep], dsRecords: [] }]);
+    expect(chain.zones[0].status).toBe('broken');
+    expect(chain.overall).toBe('broken');
+  });
+
+  it('marks an unsigned zone insecure and propagates', () => {
+    const zones: RawZone[] = [
+      { name: 'example', keys: [], dsRecords: [] },
+      { name: 'child.example', keys: [sep], dsRecords: [] },
+    ];
+    const chain = buildChain(zones);
+    expect(chain.zones[0].status).toBe('insecure');
+    expect(chain.overall).toBe('insecure');
+  });
+
+  it('marks a DS-without-DNSKEY zone as broken (not insecure)', () => {
+    const zones: RawZone[] = [
+      { name: 'example', keys: [], dsRecords: [dsFor('example')] },
+    ];
+    const chain = buildChain(zones);
+    expect(chain.zones[0].status).toBe('broken');
+    expect(chain.overall).toBe('broken');
+  });
+
+  it('marks a DS that matches no key as broken', () => {
+    const wrongDs: DsData = { ...RFC_DS, digest: Buffer.alloc(20, 0xaa) };
+    const zones: RawZone[] = [
+      { name: 'example', keys: [sep], dsRecords: [wrongDs] },
+    ];
+    const chain = buildChain(zones);
+    expect(chain.zones[0].status).toBe('broken');
+    expect(chain.overall).toBe('broken');
+  });
+
+  it('marks a zone broken when the DS key tag matches no served key', () => {
+    // Correct digest + algorithm but a deliberately wrong key tag: a validator
+    // would never select this key, so the DS authenticates nothing.
+    const good = dsFor('example');
+    const tagMismatch: DsData = { ...good, keyTag: good.keyTag + 1 };
+    const chain = buildChain([
+      { name: 'example', keys: [sep], dsRecords: [tagMismatch] },
+    ]);
+    expect(chain.zones[0].status).toBe('broken');
+  });
+
+  it('treats zones below an unsigned delegation as insecure, not broken', () => {
+    // child.example would be `broken` in isolation (its DS matches no key), but
+    // because its parent is an unsigned delegation the subtree is unauthenticated.
+    const wrongDs: DsData = { ...RFC_DS, digest: Buffer.alloc(20, 0xbb) };
+    const zones: RawZone[] = [
+      { name: 'example', keys: [], dsRecords: [] },
+      { name: 'child.example', keys: [sep], dsRecords: [wrongDs] },
+    ];
+    const chain = buildChain(zones);
+    expect(chain.zones.map((z) => z.status)).toEqual(['insecure', 'insecure']);
+    expect(chain.overall).toBe('insecure');
+  });
+
+  it('flags the root anchor and KSK metadata', () => {
+    const chain = buildChain([{ name: '.', keys: [sep], dsRecords: [] }]);
+    expect(chain.zones[0].keys[0].isSep).toBe(true);
+    expect(chain.zones[0].dsRecords).toHaveLength(ROOT_TRUST_ANCHORS.length);
+  });
+});
+
+describe('buildChain signature enforcement', () => {
+  const win = { inception: 1000, expiration: 2000 };
+  const now = 1500;
+
+  const signedZone = (name: string, algorithm: number): RawZone => {
+    const k = genKey(algorithm);
+    return {
+      name,
+      keys: [k.dnskey],
+      dsRecords: [dsForKey(name, k.dnskey)],
+      keyRrsigs: [signDnskeyRrset(name, [k.dnskey], k, win)],
+    };
+  };
+
+  it('marks a DS-linked zone secure across RSA, ECDSA, and Ed25519', () => {
+    for (const algorithm of [8, 13, 15]) {
+      const chain = buildChain([signedZone('example', algorithm)], now);
+      expect(chain.zones[0].status).toBe('secure');
+    }
+  });
+
+  it('breaks a DS-linked zone with no DNSKEY RRSIG (bad-signature)', () => {
+    const zone = signedZone('example', 13);
+    delete zone.keyRrsigs;
+    const chain = buildChain([zone], now);
+    expect(chain.zones[0].status).toBe('broken');
+    expect(chain.zones[0].breakReason).toBe('bad-signature');
+  });
+
+  it('breaks a DS-linked zone whose DNSKEY RRSIG is expired', () => {
+    const chain = buildChain([signedZone('example', 13)], win.expiration + 1);
+    expect(chain.zones[0].status).toBe('broken');
+    expect(chain.zones[0].breakReason).toBe('bad-signature');
+  });
+
+  it('breaks (not insecure) when a linked key of a supported algorithm is malformed', () => {
+    const k = genKey(13);
+    // Same DS (computed over the served, corrupted key) but the key material
+    // is unimportable: supported algorithm, broken configuration -> bogus,
+    // not an unsupported algorithm.
+    const malformed: DnskeyData = {
+      ...k.dnskey,
+      key: k.dnskey.key.subarray(0, 10),
+    };
+    const chain = buildChain(
+      [
+        {
+          name: 'example',
+          keys: [malformed],
+          dsRecords: [dsForKey('example', malformed)],
+          keyRrsigs: [signDnskeyRrset('example', [malformed], k, win)],
+        },
+      ],
+      now,
+    );
+    expect(chain.zones[0].status).toBe('broken');
+    expect(chain.zones[0].breakReason).toBe('bad-signature');
+  });
+
+  it('breaks when the DNSKEY RRSIG signer is not DS-authenticated', () => {
+    // Two keys in the RRset; a valid RRSIG by keyA, but the DS authenticates
+    // only keyB. A zone must not vouch for its own key set with an un-anchored
+    // key, so this is bogus even though the signature itself is valid.
+    const a = genKey(13);
+    const b = genKey(13);
+    const rrset = [a.dnskey, b.dnskey];
+    const chain = buildChain(
+      [
+        {
+          name: 'example',
+          keys: rrset,
+          dsRecords: [dsForKey('example', b.dnskey)], // links b, not a
+          keyRrsigs: [signDnskeyRrset('example', rrset, a, win)], // signed by a
+        },
+      ],
+      now,
+    );
+    expect(chain.zones[0].status).toBe('broken');
+    expect(chain.zones[0].breakReason).toBe('bad-signature');
+  });
+
+  it('accepts a valid RRSIG by a DS-authenticated key in a multi-key RRset', () => {
+    const a = genKey(13);
+    const b = genKey(13);
+    const rrset = [a.dnskey, b.dnskey];
+    const chain = buildChain(
+      [
+        {
+          name: 'example',
+          keys: rrset,
+          dsRecords: [dsForKey('example', a.dnskey)], // links the signer
+          keyRrsigs: [signDnskeyRrset('example', rrset, a, win)],
+        },
+      ],
+      now,
+    );
+    expect(chain.zones[0].status).toBe('secure');
+  });
+
+  it('propagates a broken key signature down the chain', () => {
+    const parent = signedZone('example', 13);
+    delete parent.keyRrsigs; // parent key set unsigned -> broken
+    const child = signedZone('child.example', 13);
+    const chain = buildChain([parent, child], now);
+    expect(chain.zones.map((z) => z.status)).toEqual(['broken', 'broken']);
+    expect(chain.overall).toBe('broken');
+  });
+});
